# avaliacao-03/src/main/java/com/avaliacao03/avaliacao03/controller/UsuariosController.java
package com.avaliacao03.avaliacao03.controller;

import java.net.URI;
import java.util.List;
import java.util.Optional;

import javax.transaction.Transactional;
import javax.validation.Valid;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.util.UriComponentsBuilder;

import com.avaliacao03.avaliacao03.config.exception.EmailExistenteException;
import com.avaliacao03.avaliacao03.controller.dto.UsuariosDto;
import com.avaliacao03.avaliacao03.controller.form.UsuarioForm;
import com.avaliacao03.avaliacao03.modelo.Usuario;
import com.avaliacao03.avaliacao03.repository.UsuarioRepository;

@RestController
@CrossOrigin()
@RequestMapping(value = "/api/v1/usuarios")
public class UsuariosController {
	
	@Autowired
	private UsuarioRepository usuarioRepository;
	
	@GetMapping
	public List<UsuariosDto> listar(){
		List<Usuario> usuarios = usuarioRepository.findAll();	
		return UsuariosDto.converter(usuarios);
	}	
	
	@PostMapping
	@Transactional
	public ResponseEntity<UsuariosDto> cadastrar(@RequestBody @Valid UsuarioForm usuarioForm,
			UriComponentsBuilder uriBuilder){
		if(usuarioRepository.findByEmail(usuarioForm.getEmail()).isPresent()) {
			throw new EmailExistenteException();
		}
		Usuario usuario = usuarioForm.converter();
		usuarioRepository.save(usuario);
		
		URI uri = uriBuilder.path("/topico/{id}").buildAndExpand(usuario.getId()).toUri();
		return ResponseEntity.created(uri).body(new UsuariosDto(usuario));	
	}
	
	@PutMapping("{id}")
	@Transactional
	public ResponseEntity<UsuariosDto> atualizar(@PathVariable Long id, @RequestBody @Valid UsuarioForm usuarioForm){
		Optional<Usuario> optional = usuarioRepository.findById(id);
		if(optional.isPresent()) {
			Usuario usuario = usuarioForm.atualizar(optional.get());
			return ResponseEntity.ok(new UsuariosDto(usuario));
		}
		return ResponseEntity.badRequest().build();
	}
}